+++ main.go
@@ -4,7 +4,6 @@ import (
 	"bufio"
 	"encoding/json"
 	"fmt"
-	"time"
 	"io/ioutil"
 	"net/http/httputil"
 	"os"
@@ -13,6 +12,7 @@ import (
 	"sort"
 	"strconv"
 	"strings"
+	"time"
 
 	"github.com/pborman/getopt"
 	"github.com/starkandwayne/goutils/ansi"
@@ -54,7 +54,41 @@ func main() {
 	go Signals()
 
 	r := NewRunner()
-	r.Dispatch("version", func(command string, args ...string) error {
+	r.HelpTopic("usage", `Usage: safe <command> [options...]
+
+Valid subcommands are:
+
+    @C{targets}     List all Vaults that have been targeted
+    @C{target}      Target a new or existing Vault
+    @C{auth}        Authenticate against the currently targeted Vault
+
+    @G{tree}        Print a tree listing of all reachable keys for each path
+
+    @G{read}        Retrieve and print the values of one or more paths
+    @R{write}       Update a single path with new keys
+
+    @R{delete}      Remove multiple paths from the Vault
+    @R{move}        Move a secret from one path to another
+    @R{copy}        Copy a secret from one path to another
+
+    @R{gen}         Generate a new, random secret
+    @R{ssh}         Generate a new SSH RSA keypair
+    @R{rsa}         Generate a new RSA keypair
+
+    @G{export}      Export one or more paths to a backup file
+    @R{import}      Import a Vault backup file
+
+    @R{vault}       Run arbitrary commands through the Vault CLI
+
+Try 'safe help <thing>' for detailed information,
+ or 'safe commands' for a list of all commands.
+`)
+
+	r.Dispatch("version", &Help{
+		Summary: "Print the version of the safe CLI",
+		Usage:   "safe version",
+		Type:    AdministrativeCommand,
+	}, func(command string, args ...string) error {
 		if Version != "" {
 			fmt.Fprintf(os.Stderr, "safe v%s\n", Version)
 		} else {
@@ -64,145 +98,28 @@ func main() {
 		return nil
 	})
 
-	r.Dispatch("help", func(command string, args ...string) error {
-		fmt.Fprintf(os.Stderr, `Usage: safe <cmd> <args ...>
-
-    Valid subcommands are:
-
-    targets
-           List all Vaults that have been targeted.
-
-    target [vault-address] name
-           Target a new or existing Vault.
-
-    auth [token|ldap|github]
-           Authenticate against the currently targeted Vault.
-
-    status
-           Print the status of all backends to the current target.
-
-    seal
-           Seal all backend Vaults for the current target.
-
-    unseal
-           Unseal all backend Vaults for the current target.
-
-    get path [path ...]
-           Retrieve and print the values of one or more paths.
-
-    set path key[=value] [key ...]
-           Update a single path with new keys.  Any existing keys that are
-           not specified on the command line are left intact. You will be
-           prompted to enter values for any keys that do not have values.
-           This can be used for more sensitive credentials like passwords,
-           PINs, etc.
-
-    paste path key[=value] [key ...]
-           Works the same way as 'safe set', except that it does not
-           prompt for confirmation of any values. This is used when you are
-           pasting in data from an external source, and do not expect to
-           mis-paste the data, to save a little time + headache.
-
-    paths path [path ... ]
-           Provide a flat listing of all reachable keys for each path.
-
-    tree [-d] path [path ...]
-           Provide a tree hierarchy listing of all reachable keys for each path.
-           The optional -d argument will hide the leaf nodes in the tree, and
-           only print the interior (directory) nodes.  This can lead to much more
-           concise output, useful when you're trying to get your bearings.
-
-    delete path [path ...]
-           Remove multiple paths from the Vault.
-
-    move oldpath newpath
-           Move a secret from oldpath to newpath, a rename of sorts.
-
-    copy oldpath newpath
-           Copy a secret from oldpath to newpath.
-
-    fmt format_type path oldkey newkey
-           Take the value found at path:oldkey, and reformat it based
-           on the provided flags (such as base64 encoding or crypt
-           hashing). The resultant value will be stored into path:newkey.
-
-           Valid format_types include the following:
-           - crypt-sha512
-           - base64
-
-    gen [length] path key
-           Generate a new, random password (length defaults to 64 chars).
-
-    ssh [nbits] path [path ...]
-           Generate a new SSH RSA keypair, adding the keys "private" and
-           "public" to each path. The public key will be encoded as an
-           authorized keys. The private key is a PEM-encoded DER private
-           key. (nbits defaults to 2048 bits)
-
-    rsa [nbits] path [path ...]
-           Generate a new RSA keypair, adding the keys "private" and "public"
-           to each path. Both keys will be PEM-encoded DER. (nbits defaults
-           to 2048 bits)
-
-    pki init
-           Configure your Vault to do PKI via the other safe PKI commands.
-           You have to run this command first, before you can use the 'cert',
-           'revoke', 'ca-pem' and 'crl-pem' commands (unless you've already
-           set up the pki backend on your Vault, in which case, cheers!)
-
-    cert role path
-           Generates a signed Certificate using Vault's PKI backend + Certifiate
-           Authority using the provided role. The common name is derived from the
-           last part of the path provided. Once issued, safe will store the
-           private key, signed certificate, and serial number in the secret backend,
-           located at the specified path.
-
-           The --ttl, --ip-sans, --alt-names, and --exclude-cn-from-sans flags can
-           be specified to customize how the certificate is generated.
-
-
-    revoke path|serial
-           Revokes a certificate from Vaults PKI backend, using the specified serial,
-           or path to a secret containing the serial of the certificate. Once revoked,
-           the CRL will be automatically updated inside Vault, but anything consuming
-           the CRL should pull a new copy.
-
-    ca-pem [path]
-           Retrieves the PEM-encoded CA cert used in Vault's PKI backend for signing
-           and issuing certificates. If path is supplied, sets the "ca-pem" key using the
-           current CA cert inside the secret backend, at <path>.
-
-    crl-pem [path]
-           Retrieves the PEM-encoded Certificate Revocation List managed by
-           Vaults PKI backend. If path is supplied, sets the "crl-pem" key using the
-           current CRL inside the secret backend, at <path>.
-
-    dhparam [bits] path
-           Generates DH Params using OpenSSL, and the specified bit length. Defaults
-           to 2048 bit primes. Primes are then stored in <path> under the 'dhparam-pem'
-           key.
-
-    prompt ...
-           Echo the arguments, space-separated, as a single line to the terminal.
-
-    import <export.file
-           Read from STDIN an export file and write all of the secrets contained
-           therein to the same paths inside the Vault
-
-    export path [path ...]
-           Export the given subtree(s) in a format suitable for migration (via a
-           future import call), or long-term storage offline.
-
-    vault  ...
-           Runs arbitrary commands through the vault cli.
-`)
+	r.Dispatch("help", nil, func(command string, args ...string) error {
+		if len(args) == 0 {
+			args = append(args, "usage")
+		}
+		if len(args) > 1 {
+			ansi.Fprintf(os.Stderr, "@R{too many arguments to `safe help'}\n")
+			r.Help(os.Stderr, "usage")
+			os.Exit(1)
+			return nil
+		}
+		r.Help(os.Stderr, args[0])
 		os.Exit(0)
 		return nil
 	})
 
-	r.Dispatch("targets", func(command string, args ...string) error {
+	r.Dispatch("targets", &Help{
+		Summary: "List all targeted Vaults",
+		Usage:   "safe targets",
+		Type:    AdministrativeCommand,
+	}, func(command string, args ...string) error {
 		if len(args) != 0 {
-			return fmt.Errorf("USAGE: targets")
+			r.ExitWithUsage("targets")
 		}
 
 		cfg := rc.Apply()
@@ -233,7 +150,11 @@ func main() {
 		return nil
 	})
 
-	r.Dispatch("target", func(command string, args ...string) error {
+	r.Dispatch("target", &Help{
+		Summary: "Target a new Vault, or set your current Vault target",
+		Usage:   "safe target [URL] [ALIAS]",
+		Type:    AdministrativeCommand,
+	}, func(command string, args ...string) error {
 		cfg := rc.Apply()
 		if len(args) == 0 {
 			if cfg.Current == "" {
@@ -266,10 +187,15 @@ func main() {
 			return cfg.Write()
 		}
 
-		return fmt.Errorf("USAGE: target [vault-address] name")
+		r.ExitWithUsage("target")
+		return nil
 	})
 
-	r.Dispatch("status", func(command string, args ...string) error {
+	r.Dispatch("status", &Help{
+		Summary: "Print the status of the current targets backend nodes",
+		Usage:   "safe status",
+		Type:    AdministrativeCommand,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 		v := connect()
 		st, err := v.Strongbox()
@@ -287,7 +213,11 @@ func main() {
 		return nil
 	})
 
-	r.Dispatch("unseal", func(command string, args ...string) error {
+	r.Dispatch("unseal", &Help{
+		Summary: "Unseal the current target",
+		Usage:   "safe unseal",
+		Type:    AdministrativeCommand,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 		v := connect()
 		st, err := v.Strongbox()
@@ -336,7 +266,11 @@ func main() {
 		return nil
 	})
 
-	r.Dispatch("seal", func(command string, args ...string) error {
+	r.Dispatch("seal", &Help{
+		Summary: "Seal the current target",
+		Usage:   "safe seal",
+		Type:    AdministrativeCommand,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 		v := connect()
 		st, err := v.Strongbox()
@@ -360,7 +294,7 @@ func main() {
 				if state == "unsealed" {
 					v.URL = addr
 
-					sealed, err := v.Seal();
+					sealed, err := v.Seal()
 					if err != nil {
 						return err
 					}
@@ -372,21 +306,29 @@ func main() {
 				}
 			}
 			if n != 0 {
-			time.Sleep(500 * time.Millisecond)
+				time.Sleep(500 * time.Millisecond)
 			}
 		}
 
 		return nil
 	})
 
-	r.Dispatch("env", func(command string, args ...string) error {
+	r.Dispatch("env", &Help{
+		Summary: "Print the VAULT_ADDR and VAULT_TOKEN for the current target",
+		Usage:   "safe env",
+		Type:    AdministrativeCommand,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 		ansi.Fprintf(os.Stderr, "  @B{VAULT_ADDR}  @G{%s}\n", os.Getenv("VAULT_ADDR"))
 		ansi.Fprintf(os.Stderr, "  @B{VAULT_TOKEN} @G{%s}\n", os.Getenv("VAULT_TOKEN"))
 		return nil
 	})
 
-	r.Dispatch("auth", func(command string, args ...string) error {
+	r.Dispatch("auth", &Help{
+		Summary: "Authenticate to the current target",
+		Usage:   "safe auth (token|github|ldap)",
+		Type:    AdministrativeCommand,
+	}, func(command string, args ...string) error {
 		cfg := rc.Apply()
 
 		method := "token"
@@ -430,10 +372,24 @@ func main() {
 
 	}, "login")
 
-	r.Dispatch("set", func(command string, args ...string) error {
+	r.Dispatch("set", &Help{
+		Summary: "Create or update a secret",
+		Usage:   "safe set PATH NAME=[VALUE] [NAME ...]",
+		Type:    DestructiveCommand,
+		Description: `
+Update a single path in the Vault with new or updated named attributes.
+Any existing name/value pairs not specified on the command-line will be
+left alone, with their original values.
+
+You will be prompted to provide (and confirm) any values that are omitted.
+This can be useful for sensitive credential like passwords and PINs, when
+you don't want the value to show up in your ~/.bash_history, or in the
+process table.
+`,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 		if len(args) < 2 {
-			return fmt.Errorf("USAGE: set path key[=value] [key ...]")
+			r.ExitWithUsage("set")
 		}
 		v := connect()
 		path, args := args[0], args[1:]
@@ -451,10 +407,24 @@ func main() {
 		return v.Write(path, s)
 	}, "write")
 
-	r.Dispatch("paste", func(command string, args ...string) error {
+	r.Dispatch("paste", &Help{
+		Summary: "Create or update a secret",
+		Usage:   "safe paste PATH NAME=[VALUE] [NAME ...]",
+		Type:    DestructiveCommand,
+		Description: `
+Works just like 'safe set', updating a single path in the Vault with new or
+updated named attributes.  Any eisting name/value pairs not specified on the
+command-line will be left alone, with their original values.
+
+You will be prompted to provide any values that are omitted, but unlike the
+'safe set' command, you will not be asked to confirm those values.  This makes
+sense when you are pasting in credentials from an external password manager
+like 1password or Lastpass.
+`,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 		if len(args) < 2 {
-			return fmt.Errorf("USAGE: paste path key[=value] [key ...]")
+			r.ExitWithUsage("paste")
 		}
 		v := connect()
 		path, args := args[0], args[1:]
@@ -472,10 +442,14 @@ func main() {
 		return v.Write(path, s)
 	})
 
-	r.Dispatch("get", func(command string, args ...string) error {
+	r.Dispatch("get", &Help{
+		Summary: "Retrieve and print the values of one or more paths",
+		Usage:   "safe get PATH [PATH ...]",
+		Type:    NonDestructiveCommand,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 		if len(args) < 1 {
-			return fmt.Errorf("USAGE: get path [path ...]")
+			r.ExitWithUsage("get")
 		}
 		v := connect()
 		for _, path := range args {
@@ -497,7 +471,17 @@ func main() {
 		return nil
 	}, "read", "cat")
 
-	r.Dispatch("tree", func(command string, args ...string) error {
+	r.Dispatch("tree", &Help{
+		Summary: "Print a tree listing of one or more paths",
+		Usage:   "safe tree [-d] [PATH ...]",
+		Type:    NonDestructiveCommand,
+		Description: `
+Walks the hierarchy of secrets stored underneath a given path, listing all
+reachable name/value pairs.  If '-d' is given, only the containing folders
+will be printed; this more concise output can be useful when you're trying
+to get your bearings.
+`,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 		opt := vault.TreeOptions{
 			UseANSI: true,
@@ -520,10 +504,14 @@ func main() {
 		return nil
 	})
 
-	r.Dispatch("paths", func(command string, args ...string) error {
+	r.Dispatch("paths", &Help{
+		Summary: "Print all of the known paths, one per line",
+		Usage:   "safe paths PATH [PATH ...]",
+		Type:    NonDestructiveCommand,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 		if len(args) < 1 {
-			return fmt.Errorf("USAGE: paths path [path ...]")
+			args = append(args, "secret")
 		}
 		v := connect()
 		for _, path := range args {
@@ -540,13 +528,17 @@ func main() {
 		return nil
 	})
 
-	r.Dispatch("delete", func(command string, args ...string) error {
+	r.Dispatch("delete", &Help{
+		Summary: "Remove one or more path from the Vault",
+		Usage:   "safe delete [-R] PATH [PATH ...]",
+		Type:    DestructiveCommand,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 
 		recurse, args := shouldRecurse(command, args...)
 
 		if len(args) < 1 {
-			return fmt.Errorf("USAGE: delete path [path ...]")
+			r.ExitWithUsage("delete")
 		}
 		v := connect()
 		for _, path := range args {
@@ -563,10 +555,14 @@ func main() {
 		return nil
 	}, "rm")
 
-	r.Dispatch("export", func(command string, args ...string) error {
+	r.Dispatch("export", &Help{
+		Summary: "Export one or more subtrees for migration / backup purposes",
+		Usage:   "safe export PATH [PATH ...]",
+		Type:    NonDestructiveCommand,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 		if len(args) < 1 {
-			return fmt.Errorf("USAGE: export path [path ...]")
+			args = append(args, "secret")
 		}
 		v := connect()
 		data := make(map[string]*vault.Secret)
@@ -593,7 +589,11 @@ func main() {
 		return nil
 	})
 
-	r.Dispatch("import", func(command string, args ...string) error {
+	r.Dispatch("import", &Help{
+		Summary: "Import name/value pairs into the current Vault",
+		Usage:   "safe import <backup/file.json",
+		Type:    DestructiveCommand,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 		b, err := ioutil.ReadAll(os.Stdin)
 		if err != nil {
@@ -616,13 +616,17 @@ func main() {
 		return nil
 	})
 
-	r.Dispatch("move", func(command string, args ...string) error {
+	r.Dispatch("move", &Help{
+		Summary: "Move a secret from one path to another",
+		Usage:   "safe move [-R] OLD-PATH NEW-PATH",
+		Type:    DestructiveCommand,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 
 		recurse, args := shouldRecurse(command, args...)
 
 		if len(args) != 2 {
-			return fmt.Errorf("USAGE: move oldpath newpath", args)
+			r.ExitWithUsage("move")
 		}
 		v := connect()
 
@@ -638,13 +642,17 @@ func main() {
 		return nil
 	}, "mv", "rename")
 
-	r.Dispatch("copy", func(command string, args ...string) error {
+	r.Dispatch("copy", &Help{
+		Summary: "Copy a secret from one path to another",
+		Usage:   "safe copy [-R] OLD-PATH NEW-PATH",
+		Type:    DestructiveCommand,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 
 		recurse, args := shouldRecurse(command, args...)
 
 		if len(args) != 2 {
-			return fmt.Errorf("USAGE: copy oldpath newpath", args)
+			r.ExitWithUsage("copy")
 		}
 		v := connect()
 
@@ -660,7 +668,14 @@ func main() {
 		return nil
 	}, "cp")
 
-	r.Dispatch("gen", func(command string, args ...string) error {
+	r.Dispatch("gen", &Help{
+		Summary: "Generate a random password",
+		Usage:   "safe gen [LENGTH] PATH KEY",
+		Type:    DestructiveCommand,
+		Description: `
+LENGTH defaults to 64 characters.
+`,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 		length := 64
 		if len(args) > 0 {
@@ -671,7 +686,7 @@ func main() {
 		}
 
 		if len(args) != 2 {
-			return fmt.Errorf("USAGE: gen [length] path key")
+			r.ExitWithUsage("gen")
 		}
 
 		v := connect()
@@ -688,7 +703,17 @@ func main() {
 		return nil
 	}, "auto")
 
-	r.Dispatch("ssh", func(command string, args ...string) error {
+	r.Dispatch("ssh", &Help{
+		Summary: "Generate one or more new SSH RSA keypair(s)",
+		Usage:   "safe ssh [NBITS] PATH [PATH ...]",
+		Type:    DestructiveCommand,
+		Description: `
+For each PATH given, a new SSH RSA public/private keypair will be generated,
+with a key strength of NBITS (which defaults to 2048).  The private keys will
+be stored under the 'private' name, as a PEM-encoded RSA private key, and the
+public key, formatted for use in an SSH authorized_keys file, under 'public'.
+`,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 		bits := 2048
 		if len(args) > 0 {
@@ -699,7 +724,7 @@ func main() {
 		}
 
 		if len(args) < 1 {
-			return fmt.Errorf("USAGE: ssh [bits] path [path ...]")
+			r.ExitWithUsage("ssh")
 		}
 
 		v := connect()
@@ -718,7 +743,17 @@ func main() {
 		return nil
 	})
 
-	r.Dispatch("rsa", func(command string, args ...string) error {
+	r.Dispatch("rsa", &Help{
+		Summary: "Generate a new RSA keypair",
+		Usage:   "safe rsa [NBITS] PATH [PATH ...]",
+		Type:    DestructiveCommand,
+		Description: `
+For each PATH given, a new RSA public/private keypair will be generated with a,
+key strength of NBITS (which defaults to 2048).  The private keys will be stored
+under the 'private' name, and the public key under the 'public' name.  Both will
+be PEM-encoded.
+`,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 		bits := 2048
 		if len(args) > 0 {
@@ -729,7 +764,7 @@ func main() {
 		}
 
 		if len(args) < 1 {
-			return fmt.Errorf("USAGE: rsa [bits] path [path ...]")
+			r.ExitWithUsage("rsa")
 		}
 
 		v := connect()
@@ -748,7 +783,14 @@ func main() {
 		return nil
 	})
 
-	r.Dispatch("dhparam", func(command string, args ...string) error {
+	r.Dispatch("dhparam", &Help{
+		Summary: "Generate Diffie-Helman key exchange parameters",
+		Usage:   "safe dhparam [NBITS] PATH",
+		Type:    DestructiveCommand,
+		Description: `
+NBITS defaults to 2048.
+`,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 		bits := 2048
 
@@ -760,7 +802,7 @@ func main() {
 		}
 
 		if len(args) < 1 {
-			return fmt.Errorf("USAGE: dhparam [bits] path")
+			r.ExitWithUsage("dhparam")
 		}
 
 		path := args[0]
@@ -775,12 +817,20 @@ func main() {
 		return v.Write(path, s)
 	}, "dh", "dhparams")
 
-	r.Dispatch("prompt", func(command string, args ...string) error {
+	r.Dispatch("prompt", &Help{
+		Summary: "Print a prompt (useful for scripting safe command sets)",
+		Usage:   "safe echo Your Message Here:",
+		Type:    NonDestructiveCommand,
+	}, func(command string, args ...string) error {
 		fmt.Fprintf(os.Stderr, "%s\n", strings.Join(args, " "))
 		return nil
 	})
 
-	r.Dispatch("vault", func(command string, args ...string) error {
+	r.Dispatch("vault", &Help{
+		Summary: "Run arbitrary Vault CLI commands against the current target",
+		Usage:   "safe vault ...",
+		Type:    DestructiveCommand,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 
 		cmd := exec.Command("vault", args...)
@@ -795,11 +845,27 @@ func main() {
 		return nil
 	})
 
-	r.Dispatch("fmt", func(command string, args ...string) error {
+	r.Dispatch("fmt", &Help{
+		Summary: "Reformat an existing name/value pair, into a new name",
+		Usage:   "safe fmt FORMAT PATH OLD-NAME NEW-NAME",
+		Type:    DestructiveCommand,
+		Description: `
+Take the value stored at PATH/OLD-NAME, format it a different way, and
+then save it at PATH/NEW-NAME.  This can be useful for generating a new
+password (via 'safe gen') and then crypt'ing it for use in /etc/shadow,
+using the 'crypt-sha512' format.
+
+Supported formats:
+
+    base64          Base64 encodes the value
+    crypt-sha512    Salt and hash the value, using SHA-512, in crypt format.
+
+`,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 
 		if len(args) != 4 {
-			return fmt.Errorf("USAGE: fmt format_type path oldkey newkey")
+			r.ExitWithUsage("fmt")
 		}
 
 		fmtType := args[0]
@@ -822,9 +888,20 @@ func main() {
 		return v.Write(path, s)
 	})
 
-	r.Dispatch("pki", func(command string, args ...string) error {
+	r.Dispatch("pki", &Help{
+		Summary: "Configure the PKI backend on the target Vault",
+		Usage:   "safe pki init",
+		Type:    DestructiveCommand,
+		Description: `
+Configure your Vault to do PKI via the other safe PKI commands.
+
+You have to run this command first, before you can use the 'cert',
+'revoke', 'ca-pem' and 'crl-pem' commands (unless you've already set
+up the pki backend on your Vault, in which case, cheers!)
+`,
+	}, func(command string, args ...string) error {
 		if len(args) != 1 || args[0] != "init" {
-			return fmt.Errorf("USAGE: pki setup")
+			r.ExitWithUsage("pki")
 		}
 
 		rc.Apply()
@@ -910,7 +987,23 @@ func main() {
 		return nil
 	})
 
-	r.Dispatch("crl-pem", func(command string, args ...string) error {
+	r.Dispatch("crl-pem", &Help{
+		Summary: "Retrieve the Vault Certificate Revocation List",
+		Usage:   "safe crl-pem [PATH]",
+		Type:    NonDestructiveCommand,
+		Description: `
+@M{(You must run 'safe pki init' before you can use this command)}
+
+Retrieve the Certificate Revocation List (CRL) from the Vault PKI backend.
+This list identifies which of the certificates that Vault has issued have
+since been revoked, and should not be trusted, despite their validity.
+
+The CRL will be printed to standard output, as a PEM-encoded value.
+
+If you supply a PATH, the CRL will not be printed, but will be saved at that
+path, under the name 'crl-pem'.
+`,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 
 		v := connect()
@@ -941,7 +1034,24 @@ func main() {
 		return nil
 	})
 
-	r.Dispatch("ca-pem", func(command string, args ...string) error {
+	r.Dispatch("ca-pem", &Help{
+		Summary: "Retrieve the Vault Certificate Authority (CA) Certificate",
+		Usage:   "safe ca-pem [PATH]",
+		Type:    NonDestructiveCommand,
+		Description: `
+@M{(You must run 'safe pki init' before you can use this command)}
+
+Retrieve the Certificate Authority (CA) certificate from the Vault PKI
+backend.  To take full advantage of Vault-issued certificates, you will
+need to install the CA certificate in your trusted certificates bundle.
+
+The CA certificate will be printed to standard output, as a PEM-encoded
+value.
+
+If you supply a PATH, the CA certificate will not be printed, but will be
+saved at that path, under the name 'ca-pem'.
+`,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 
 		v := connect()
@@ -976,7 +1086,42 @@ func main() {
 		return nil
 	})
 
-	r.Dispatch("cert", func(command string, args ...string) error {
+	r.Dispatch("cert", &Help{
+		Summary: "Issue a Certificate using the Vault PKI backend",
+		Usage:   "safe cert [OPTIONS] ROLE PATH",
+		Type:    DestructiveCommand,
+		Description: `
+@M{(You must run 'safe pki init' before you can use this command)}
+
+Generate a new private key, and then issue a certificate, signed by the
+Vault Certificate Authority (CA).  The common name of the new certificate
+will be based on the last part of the provided PATH, so if you want to
+issue a certificate for secure.example.com, you want to use something like
+
+    safe cert example.com secret/certs/secure.example.com
+
+(Assuming you set up 'example.com' as your CA FQDN when you went through
+ the 'safe pki init' setup stage.)
+
+The following options are recognized:
+
+  --ttl                    How long the cert should be valid for
+                           (i.e. '90d', '10h', etc.)
+
+  --alt-names              A comma-separated list of alternate DNS
+                           names (SANs) to include in the certificate.
+
+  --ip-sans                Comma-separated list of IP addresses to
+                           include in the certificate as IP SANs
+
+  --exclude-cn-from-sans   Exclude the certificate's common name (CN)
+                           from the Subject Alternate Name list.
+
+Once generated, the new private key will be stored under the name 'key',
+the certificate will be under 'cert', a combined PEM containing both will
+be saved as 'combined', and the certificate serial number under 'serial'.
+`,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 
 		ttl := getopt.StringLong("ttl", 0, "", "Vault-compatible time specification for the length the Cert is valid for")
@@ -1007,7 +1152,7 @@ func main() {
 		}
 
 		if len(args) != 2 {
-			return fmt.Errorf("USAGE: cert role path")
+			r.ExitWithUsage("cert")
 		}
 
 		v := connect()
@@ -1019,11 +1164,27 @@ func main() {
 		return v.CreateSignedCertificate(role, path, params)
 	})
 
-	r.Dispatch("revoke", func(command string, args ...string) error {
+	r.Dispatch("revoke", &Help{
+		Summary: "Revoke a Vault-issued Certificate",
+		Usage:   "safe revoke [PATH | SERIAL]",
+		Type:    DestructiveCommand,
+		Description: `
+@M{(You must run 'safe pki init' before you can use this command)}
+
+Revoke a certificate that was issued by the Vault PKI backend.  You can
+specify which certificate to revoke by using its serial number, or by
+providing the PATH in the Vault where the certificate is stored.
+
+After the certificate is revoked, it will be added to the Certificate
+Revocation List (CRL) automatically.  Note that this will only update the
+Vault's copy of the CRL -- any external copies of the the list will need
+to be refreshed via the 'safe crl-pem' command.
+`,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 
 		if len(args) != 1 {
-			return fmt.Errorf("USAGE: revoke path|serial")
+			r.ExitWithUsage("revoke")
 		}
 
 		v := connect()
@@ -1034,11 +1195,29 @@ func main() {
 		return v.RevokeCertificate(args[0])
 	})
 
-	r.Dispatch("curl", func(command string, args ...string) error {
+	r.Dispatch("curl", &Help{
+		Summary: "Issue arbitrary HTTP requests to the current Vault (for diagnostics)",
+		Usage:   "safe curl [OPTIONS] METHOD REL-URI [DATA]",
+		Type:    DestructiveCommand,
+		Description: `
+This is a debugging and diagnostics tool.  You should not need to use
+'safe curl' for normal operation or interaction with a Vault.
+
+METHOD must be one of GET, POST, or PUT.
+
+REL-URI is the relative URI (the path component, starting with the first
+forward slash) of the resource you wish to access.
+
+DATA should be a JSON string, since almost all of the Vault API handlers
+deal exclusively in JSON payloads.  GET requests should not have DATA.
+Query string parameters should be appended to REL-URI, instead of being
+sent as DATA.
+`,
+	}, func(command string, args ...string) error {
 		rc.Apply()
 
 		if len(args) < 2 {
-			return fmt.Errorf("USAGE: curl method path [data]")
+			r.ExitWithUsage("curl")
 		}
 
 		v := connect()
@@ -1070,6 +1249,10 @@ func main() {
 
 	} else {
 		args = opts.Args()
+	}
+
+	if len(args) == 1 && args[0] == "commands" {
+		args = []string{"help", "commands"}
 	}
 
 	if *insecure {